fix: overflows no longer occur.
When the last todo on the screen is scrolled to, if there are more
todos, the interface will automatically scroll down and pin the bottom
of the previous screen at the top of the current.
New additions, when there are more todos than lines, cause
the screen to scroll down to them.

closes #4